+++ crates/apollo-compiler/src/lib.rs
@@ -18,7 +18,7 @@ use crate::database::{InputDatabase, ReprDatabase, Source};
 use crate::validation::ValidationDatabase;
 
 pub use self::executable::ExecutableDocument;
-pub use self::node::Node;
+pub use self::node::{FileId, Node, NodeLocation};
 pub use self::node_str::NodeStr;
 pub use self::parser::{parse_mixed_validate, Parser, SourceFile, SourceMap};
 pub use self::schema::Schema;